api/requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.8.0
httpx>=0.25.2
python-multipart>=0.0.6
psycopg2-binary>=2.9.0
python-dotenv>=1.0.0
mangum>=0.17.0
supabase>=2.0.0

